sdk/requirements.txt
# 核心依赖
pyinstrument>=4.6.0
requests>=2.25.0
pyyaml>=5.4.0
psutil>=5.8.0

# 异步支持
aiohttp>=3.8.0

# 数据序列化
msgpack>=1.0.0

# 可选框架依赖（通过extras_require安装）
# Flask>=1.0.0
# Django>=2.2.0  
# fastapi>=0.68.0

# 开发依赖
# pytest>=6.0.0
# pytest-cov>=2.10.0
# pytest-asyncio>=0.18.0
# black>=21.0.0
# isort>=5.0.0
# flake8>=3.8.0
# mypy>=0.800